Modules and ES6 implemented in the code
