+++ index.html
@@ -5,7 +5,7 @@
 		<meta http-equiv="X-UA-Compatible" content="IE=edge" />
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<link rel="stylesheet" href="./style.css" />
-		<script src=./index.js defer> </script>
+
 		<title>Awesome books</title>
 	</head>
 	<body>
@@ -47,5 +47,6 @@
 			</div>
 		</section>
 		<footer id="footer"><span>copyright(2022)</span><span>...</span></footer> 
+		<script type="module" src=./index.js defer> </script>
 	</body> 
 </html>
