+++ index.js
@@ -1,8 +1,11 @@
+import './module/dateSet.js';
+import BookApp, { sectionBooks } from './module/classUpdate.js';
 const inputTitle = document.querySelector('.title');
 const inputAuthor = document.querySelector('.author');
 const addBtn = document.querySelector('.add-btn');
-const sectionBooks = document.querySelector('.books');
 const footer = document.getElementById('footer');
+
+const books = new BookApp();
 
 const findHight = () => {
   if (window.innerHeight > document.body.scrollHeight) {
@@ -12,7 +15,7 @@ const findHight = () => {
   }
 };
 
-function loadBooks(bookCollection) {
+export default function loadBooks(bookCollection) {
   sectionBooks.innerHTML = '';
   /* eslint-disable */
   if (bookCollection.length > 0) {
@@ -22,66 +25,10 @@ function loadBooks(bookCollection) {
         books.returnBook(book.title, book.author, i)
       );
     });
-    console.log(window.innerHeight);
-    console.log(document.body.scrollHeight);
   }
   books.colorGiver();
   /* eslint-enable */
 }
-
-class BookApp {
-  height = 0;
-
-  constructor() {
-    /* eslint-disable */
-    this.bookCollection = localStorage.getItem('bookCollection') ?
-      JSON.parse(localStorage.getItem('bookCollection')) : [];
-    /* eslint-enable */
-  }
-
-  addBook(title, author) {
-    this.bookCollection.push({
-      title,
-      author,
-    });
-    localStorage.setItem('bookCollection', JSON.stringify(this.bookCollection));
-    loadBooks(this.bookCollection);
-  }
-
-  removeBook(i) {
-    this.bookCollection.splice(i, 1);
-    localStorage.setItem('bookCollection', JSON.stringify(this.bookCollection));
-    loadBooks(this.bookCollection);
-  }
-  /* eslint-disable */
-  returnBook(title, author, i) {
-    const html = `<div class="each-book">
-      <div class="book-align">
-		  <span class="book-title">"${title}" by </span>
-		  <span class="author">${author}</span>
-      </div>
-		  <button class="remove-btn" data-id=${i}>Remove</button>
-	  </div>`;
-    return html;
-  }
-
-  /* eslint-enable */
-  colorGiver() {
-    if (this.bookCollection.length !== 0) {
-      sectionBooks.style.border = '2px solid black';
-      const divEl = document.querySelectorAll('.each-book');
-      divEl.forEach((div, i) => {
-        if (i % 2 === 0) div.style.background = 'white';
-        else div.style.background = 'lightgray';
-      });
-    }
-    if (this.bookCollection.length === 0) {
-      sectionBooks.style.border = 'none';
-    }
-  }
-}
-
-const books = new BookApp();
 
 addBtn.addEventListener('click', (e) => {
   e.preventDefault();
@@ -116,26 +63,4 @@ lists.forEach((list) => {
     document.querySelector(`.${target}`).classList.remove('hidden');
     findHight();
   });
-});
-
-const dateEl = document.querySelector('.date');
-
-const setDate = function () {
-  const now = new Date();
-  const options = {
-    month: 'long',
-    day: '2-digit',
-    year: 'numeric',
-    hour: '2-digit',
-    minute: '2-digit',
-    second: '2-digit',
-    hour12: true,
-  };
-  const date = Intl.DateTimeFormat([], options).format(now);
-  return date;
-};
-
-setInterval(() => {
-  dateEl.textContent = '';
-  dateEl.textContent = setDate();
 });
